+++ cashu/wallet/cli/cli.py
