+++ tests/test_cli.py
@@ -7,7 +7,7 @@ from click.testing import CliRunner
 from cashu.core.migrations import migrate_databases
 from cashu.core.settings import VERSION
 from cashu.wallet import migrations
-from cashu.wallet.cli import cli
+from cashu.wallet.cli.cli import cli
 from cashu.wallet.wallet import Wallet
 from tests.conftest import SERVER_ENDPOINT, mint
